Add `recordsink` package and move stuff into it.
