+++ service/src/main/java/org/databiosphere/workspacedataservice/service/BatchWriteServiceConfig.java
@@ -1,7 +1,8 @@
 package org.databiosphere.workspacedataservice.service;
 
 import org.databiosphere.workspacedataservice.config.TwdsProperties;
-import org.databiosphere.workspacedataservice.service.BatchWriteService.RecordSink;
+import org.databiosphere.workspacedataservice.recordsink.RawlsRecordSink;
+import org.databiosphere.workspacedataservice.recordsink.RecordSink;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
